get reports

+++ database/migrations/2019_01_27_051357_delete_status_in_transaction_flows_table.php
@@ -14,6 +14,7 @@ class DeleteStatusInTransactionFlowsTable extends Migration
     public function up()
     {
         Schema::table('transaction_flows', function (Blueprint $table) {
+            if(!Schema::hasColumn('transaction_flows', 'status')) return false;
             $table->dropColumn('status');
         });
     }
